+++ src/search/resilient_node.h
@@ -16,8 +16,6 @@ class ResilientNode
 
 public:
     ResilientNode(State state_, int k_, std::set<Operator> deactivated_op_);
-    //ResilientNode(State state_, int k_);
-    //ResilientNode(State state_);
 
     void dump() const;
 
@@ -31,7 +29,7 @@ public:
         return k;
     }
 
-    size_t get_id() const
+    int get_id() const
     {
         return id;
     }
